--- src/features/todo/api/mocks/todolistMockData.ts
@@ -1,62 +1,0 @@
-export const todolist = {
-  todolistId: '123123123123',
-  todolist: [
-    {
-      id: '0000',
-      content: '튜토리얼 영상 1-3 시청',
-      createdAt: new Date(),
-      completedAt: new Date(),
-      completed: false,
-      shared: true,
-    },
-    {
-      id: '1111',
-      content: '주요 단축키 정리 노션에 업로드',
-      createdAt: new Date(),
-      completedAt: new Date(),
-      completed: false,
-      shared: true,
-    },
-    {
-      id: '2222',
-      content: '키 인사이트 3개 도출해서 슬랙에 공유',
-      createdAt: new Date(),
-      completedAt: new Date(),
-      completed: false,
-      shared: true,
-    },
-    {
-      id: '3333',
-      content: '간단한 모바일 UI 따라 그리기',
-      createdAt: new Date(),
-      completedAt: new Date(),
-      completed: false,
-      shared: true,
-    },
-    {
-      id: '4444',
-      content: '1번 개인 할일',
-      createdAt: new Date(),
-      completedAt: new Date(),
-      completed: false,
-      shared: false,
-    },
-    {
-      id: '5555',
-      content: '2번 개인 할일',
-      createdAt: new Date(),
-      completedAt: new Date(),
-      completed: false,
-      shared: false,
-    },
-    {
-      id: '6666',
-      content: '3번 개인 할일',
-      createdAt: new Date(),
-      completedAt: new Date(),
-      completed: false,
-      shared: false,
-    },
-  ],
-  order: ['0000', '1111', '2222', '3333', '4444', '5555', '6666'],
-};
