+++ src/features/todo/api/todolist.ts
@@ -1,23 +1,5 @@
 import { clientFetch } from '@/shared/api';
 
-// export const fetchTodolist = async (
-//   todolistId: string,
-//   server: boolean = false,
-// ) => {
-//   let url = `/api/todolist/${todolistId}`;
-//   if (server) url = `${process.env.NEXT_PUBLIC_MOCKSERVER_URL}` + url;
-//   else url = `${process.env.NEXT_PUBLIC_BASE_URL}` + url;
-//   console.log('fetch');
-
-//   const response = await fetch(url);
-
-//   if (!response.ok) {
-//     throw new Error('Failed to fetch todolist');
-//   }
-
-//   return await response.json();
-// };
-
 export const fetchTodolist = async (todolistId: string) => {
   const endpoint = `/api/todolist/${todolistId}`;
   const parsedResponse = await clientFetch.get(endpoint);
@@ -25,16 +7,16 @@ export const fetchTodolist = async (todolistId: string) => {
   return parsedResponse;
 };
 
-export interface newTodoData {
-  content: string;
-  shared: boolean;
-}
-export const createTodo = async (todolistId: string, newTodo: newTodoData) => {
-  const endpoint = `/api/todolist/${todolistId}/todo`;
-  const parsedResponse = await clientFetch.post(endpoint, newTodo);
+// export interface newTodoData {
+//   content: string;
+//   shared: boolean;
+// }
+// export const createTodo = async (todolistId: string, newTodo: newTodoData) => {
+//   const endpoint = `/api/todolist/${todolistId}/todo`;
+//   const parsedResponse = await clientFetch.post(endpoint, newTodo);
 
-  return parsedResponse;
-};
+//   return parsedResponse;
+// };
 
 export interface newContent {
   content: string;
